# fleet_management_api/openapi/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: fleet@bringauto.com
    name: BringAuto s.r.o
    url: https://bringauto.com
  description: Specification for BringAuto fleet backend HTTP API
  license:
    name: AGPLv3
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  title: BringAuto Fleet Management v2 API
  version: 3.0.0
servers:
- url: /v2/management
security:
- APIKeyAuth: []
- oAuth2AuthCode: []
...
paths:
...
  /car:
    get:
      operationId: get_cars
      responses:
        "200":
          content:
            application/json:
              example:
              - id: 1
                platformHWId: 5
                name: BAT-2022-01
                carAdminPhone:
                  phone: "+420123456789"
                defaultRouteId: 1
                underTest: false
              schema:
                items:
                  $ref: '#/components/schemas/Car'
                type: array
          description: All the currently existing Cars have been returned.
        "401":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "401"
                    message: Unauthorized
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "403"
                    message: Request forbidden
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        default:
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "500"
                    message: Unexpected error
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Find and return all existing Cars.
      tags:
      - car
      x-openapi-router-controller: fleet_management_api.api_impl.controllers.car
    post:
      operationId: create_cars
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/Car'
              type: array
        description: A Car model in JSON format.
        required: true
        x-body-name: car
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Car'
                type: array
          description: The Car has been successfully created.
        "400":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "400"
                    message: Bad request
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "401":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "401"
                    message: Unauthorized
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "403"
                    message: Request forbidden
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        default:
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "500"
                    message: Unexpected error
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Create a new Car object.
      tags:
      - car
      x-openapi-router-controller: fleet_management_api.api_impl.controllers.car
    put:
      operationId: update_cars
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/Car'
              type: array
        description: JSON representation of the updated Car.
        required: true
        x-body-name: car
      responses:
        "200":
          content:
            text/plain: {}
          description: The Car has been successfully updated.
        "400":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "400"
                    message: Bad request
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "401":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "401"
                    message: Unauthorized
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "403"
                    message: Request forbidden
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "404"
                    message: Resource not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
        default:
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "500"
                    message: Unexpected error
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Update already existing Car.
      tags:
      - car
      x-openapi-router-controller: fleet_management_api.api_impl.controllers.car
  /car/{carId}:
    delete:
      operationId: delete_car
      parameters:
      - description: The ID of the Car to be deleted.
        example: 1
        in: path
        name: carId
        required: true
        schema:
          format: int32
          type: integer
      responses:
        "200":
          content:
            text/plain:
              example: 2 cars were successfully removed.
          description: The Car has been successfully deleted.
        "400":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "400"
                    message: Bad request
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "401":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "401"
                    message: Unauthorized
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "403"
                    message: Request forbidden
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "404"
                    message: Resource not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
        "405":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "405"
                    message: Method not allowed
              schema:
                $ref: '#/components/schemas/Error'
          description: Method not allowed
        default:
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "500"
                    message: Unexpected error
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Delete a Car identified by its ID.
      tags:
      - car
      x-openapi-router-controller: fleet_management_api.api_impl.controllers.car
    get:
      operationId: get_car
      parameters:
      - description: An ID of the Car to be returned.
        example: 1
        in: path
        name: carId
        required: true
        schema:
          format: int32
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
          description: The Car with the specified ID has been found and returned.
        "400":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "400"
                    message: Bad request
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "401":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "401"
                    message: Unauthorized
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "403"
                    message: Request forbidden
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "404"
                    message: Resource not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
        default:
          content:
            application/json:
              examples:
                error:
                  value:
                    code: "500"
                    message: Unexpected error
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Find a single Car by its ID.
      tags:
      - car
      x-openapi-router-controller: fleet_management_api.api_impl.controllers.car
...
components:
...
  schemas:
    Error:
      description: Error object structure.
      properties:
        code:
          example: 404
          title: code
          type: integer
        message:
          example: Not found
          title: message
          type: string
      required:
      - code
      - message
      title: Error
      type: object
    Car:
      description: Car model structure.
      example:
        underTest: false
        name: BAT-2022-01
        defaultRouteId: 1
        id: 1
        carAdminPhone:
          phone: "+420123456789"
        lastState:
          fuel: 80
          id: 1
          position:
            altitude: 400.25
            latitude: 49.204117
            longitude: 16.606525
          speed: 20.5
          timestamp: 1616425275913
          status: idle
          carId: 1
        platformHwId: 1
      properties:
        id:
          example: 1
          format: int32
          title: Id
          type: integer
        platformHwId:
          example: 1
          format: int32
          title: Id
          type: integer
        name:
          example: BAT-2022-01
          title: name
          type: string
        carAdminPhone:
          $ref: '#/components/schemas/MobilePhone'
        defaultRouteId:
          example: 1
          format: int32
          title: Id
          type: integer
        underTest:
          default: true
          example: false
          title: underTest
          type: boolean
        lastState:
          $ref: '#/components/schemas/CarState'
      required:
      - carAdminPhone
      - name
      - platformHwId
      title: Car
      type: object
    CarState:
      description: Car State object structure
      example:
        fuel: 80
        id: 1
        position:
          altitude: 400.25
          latitude: 49.204117
          longitude: 16.606525
        speed: 20.5
        timestamp: 1616425275913
        status: idle
        carId: 1
      properties:
        id:
          example: 1
          format: int32
          title: Id
          type: integer
        timestamp:
          description: A Unix timestamp in milliseconds. The timestamp is used to
            determine the time of creation of an object.
          example: 1616425275913
          format: int64
          title: Timestamp
          type: integer
        status:
          $ref: '#/components/schemas/CarStatus'
        fuel:
          default: 0
          example: 80
          format: int32
          title: fuel
          type: integer
        carId:
          example: 1
          format: int32
          title: Id
          type: integer
        speed:
          default: 0.0
          example: 20.5
          format: float
          title: speed
          type: number
        position:
          $ref: '#/components/schemas/GNSSPosition'
      required:
      - carId
      - status
      title: CarState
      type: object
    CarStatus:
      description: Car Status enum
      enum:
      - idle
      - charging
      - out_of_order
      - driving
      - in_stop
      - paused_by_phone
      - paused_by_obstacle
      - paused_by_button
      example: idle
      title: CarStatus
      type: string
...
    MobilePhone:
      description: MobilePhone Primitive structure.
      example:
        phone: "+420123456789"
      properties:
        phone:
          example: "+420123456789"
          title: phone
          type: string
      title: MobilePhone
      type: object
...
    GNSSPosition:
      description: GNSSPosition primitive structure.
      example:
        altitude: 400.25
        latitude: 49.204117
        longitude: 16.606525
      properties:
        latitude:
          default: 0.0
          example: 49.204117
          format: float
          title: latitude
          type: number
        longitude:
          default: 0.0
          example: 16.606525
          format: float
          title: longitude
          type: number
        altitude:
          default: 0.0
          example: 400.25
          format: float
          title: altitude
          type: number
      title: GNSSPosition
      type: object
  securitySchemes:
    APIKeyAuth:
      description: Authentication using an API key.
      in: query
      name: api_key
      type: apiKey
      x-apikeyInfoFunc: fleet_management_api.controllers.security_controller.info_from_APIKeyAuth
    oAuth2AuthCode:
      description: oAuth2 authorization
      flows:
        authorizationCode:
          authorizationUrl: https://keycloak.bringauto.com/realms/bringauto/protocol/openid-connect/auth
          refreshUrl: https://keycloak.bringauto.com/realms/bringauto/protocol/openid-connect/token
          scopes: {}
          tokenUrl: https://keycloak.bringauto.com/realms/bringauto/protocol/openid-connect/token
      type: oauth2
      x-tokenInfoFunc: fleet_management_api.controllers.security_controller.info_from_oAuth2AuthCode
      x-scopeValidateFunc: fleet_management_api.controllers.security_controller.validate_scope_oAuth2AuthCode